move IObjlistentry to engine
