+++ datafiles/javascript/engine/objects/IObjlistentry.js
@@ -1,5 +1,3 @@
-// TODO move to engine
-
 // All objects that can be stored in room.objects have to extend from this so
 // that they are guaranteed to have the methods.
